LayoutTests/fast/dom/Text/splitText-expected.txt

Tests the splitText API arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.data is "abcdefg"
PASS text.splitText(4).data is "efg"
PASS text.data is "abcd"
PASS text.splitText() threw exception TypeError: Not enough arguments.
PASS text.splitText(999) threw exception Error: IndexSizeError: DOM Exception 1.
PASS text.splitText(-1) threw exception Error: IndexSizeError: DOM Exception 1.
PASS text.data is "abcd"
PASS text.splitText(-4294967294).data is "cd"
PASS text.data is "ab"
PASS successfullyParsed is true

TEST COMPLETE
